feat: 8.23: Subscriptions - server.
Refactored all the Server code.
