+++ part08/books_graphql/package.json
@@ -12,12 +12,17 @@
   "license": "ISC",
   "dependencies": {
     "@apollo/server": "^4.2.2",
+    "@graphql-tools/schema": "^9.0.17",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.3",
+    "express": "^4.18.2",
     "graphql": "^16.6.0",
     "graphql-tag": "^2.12.6",
+    "graphql-ws": "^5.12.0",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^6.8.1",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "ws": "^8.12.1"
   },
   "devDependencies": {
     "cross-env": "^7.0.3",
